feat: continuando a implementação da funcionalidade de adicionar os papéis e permissões aos usuários

+++ routes/web.php
@@ -18,34 +18,34 @@ Route::get('/', function () {
 
 // Rotas restritas
 Route::group(['middleware' => 'auth'], function () {
-    
+
     // Perfil
-    Route::get('/profile/{user}', [DashboardController::class, 'profile'])->name('profile');
-    Route::get('/profile/{user}/edit', [DashboardController::class, 'edit'])->name('profile.edit');
-    Route::put('/profile/{user}', [DashboardController::class, 'update'])->name('profile.update');
+    Route::get('/profile/{user}', [DashboardController::class, 'profile'])->name('profile')->middleware('permission:profile');
+    Route::get('/profile/{user}/edit', [DashboardController::class, 'edit'])->name('profile.edit')->middleware('permission:profile.edit');
+    Route::put('/profile/{user}', [DashboardController::class, 'update'])->name('profile.update')->middleware('permission:profile.update');
 
     // Dashboard
-    Route::get('/dashboard', [DashboardController::class, 'index'])->name('dashboard');
+    Route::get('/dashboard', [DashboardController::class, 'index'])->name('dashboard')->middleware('permission:dashboard');
 
     // Histórico de edições
-    Route::get('/edited-records/{table}/{register}', [EditedRecordsController::class, 'index'])->name('edited.records');
+    Route::get('/edited-records/{table}/{register}', [EditedRecordsController::class, 'index'])->name('edited.records')->middleware('permission:edited.records');
 
     // Usuários
     Route::prefix('users')->group(function () {
-        Route::get('/', [UserController::class, 'index'])->name('users.index'); // Listar registros
-        Route::post('/info-create', [UserController::class, 'infoCreate'])->name('users.info-create');
-        Route::match(['get', 'post'], '/create', [UserController::class, 'create'])->name('users.create'); // Carregar formulário cadastrar registro
-        Route::get('/select-enterprise', [UserController::class, 'selectEnterprise'])->name('users.select-enterprise');
-        Route::post('/select-enterprise', [UserController::class, 'selectEnterpriseActive'])->name('users.select-enterprise-active');
-        Route::post('/', [UserController::class, 'store'])->name('users.store'); // Cadastrar registro
-        Route::get('/{user}', [UserController::class, 'show'])->name('users.show'); // Vizualizar detalhes do registro
-        Route::get('/{user}/edit', [UserController::class, 'edit'])->name('users.edit'); // Carregar formulário que edita o registro
-        Route::get('/select-enterprise-update/{user}', [UserController::class, 'selectEnterpriseUpdate'])->name('users.select-enterprise-update');
-        Route::post('/select-enterprise-update/{user}', [UserController::class, 'selectEnterpriseActiveUpdate'])->name('users.select-active-update');
-        Route::put('/{user}', [UserController::class, 'update'])->name('users.update'); // Editar o registro
-        Route::get('/{user}/edit-password', [UserController::class, 'editPassword'])->name('users.edit-password'); // Carrega o formulário que edita a senha
-        Route::put('/{user}/update-password', [UserController::class, 'updatePassword'])->name('users.update-password'); // Edita a senha
-        Route::delete('/{user}', [UserController::class, 'destroy'])->name('users.destroy')->middleware(); // Exclui o registro
+        Route::get('/', [UserController::class, 'index'])->name('users.index')->middleware('permission:users.index'); // Listar registros
+        Route::post('/info-create', [UserController::class, 'infoCreate'])->name('users.info-create')->middleware('permission:users.info-create');
+        Route::match(['get', 'post'], '/create', [UserController::class, 'create'])->name('users.create')->middleware('permission:users.create'); // Carregar formulário cadastrar registro
+        Route::get('/select-enterprise', [UserController::class, 'selectEnterprise'])->name('users.select-enterprise')->middleware('permission:users.select-enterprise');
+        Route::post('/select-enterprise', [UserController::class, 'selectEnterpriseActive'])->name('users.select-enterprise-active')->middleware('permission:users.select-enterprise-active');
+        Route::post('/', [UserController::class, 'store'])->name('users.store')->middleware('permission:users.store'); // Cadastrar registro
+        Route::get('/{user}', [UserController::class, 'show'])->name('users.show')->middleware('permission:users.show'); // Vizualizar detalhes do registro
+        Route::get('/{user}/edit', [UserController::class, 'edit'])->name('users.edit')->middleware('permission:users.edit'); // Carregar formulário que edita o registro
+        Route::get('/select-enterprise-update/{user}', [UserController::class, 'selectEnterpriseUpdate'])->name('users.select-enterprise-update')->middleware('permission:users.select-enterprise-update');
+        Route::post('/select-enterprise-update/{user}', [UserController::class, 'selectEnterpriseActiveUpdate'])->name('users.select-active-update')->middleware('permission:users.select-active-update');
+        Route::put('/{user}', [UserController::class, 'update'])->name('users.update')->middleware('permission:users.update'); // Editar o registro
+        Route::get('/{user}/edit-password', [UserController::class, 'editPassword'])->name('users.edit-password')->middleware('permission:users.edit-password'); // Carrega o formulário que edita a senha
+        Route::put('/{user}/update-password', [UserController::class, 'updatePassword'])->name('users.update-password')->middleware('permission:users.update-password'); // Edita a senha
+        Route::delete('/{user}', [UserController::class, 'destroy'])->name('users.destroy')->middleware('permission:users.destroy'); // Exclui o registro
     });
 
     // Empresas, produtos, planos, filiais e níveis de acesso
